requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.12.1
diff-match-patch==20230430
Django==5.0.6
django-bootstrap-datepicker==1.5.0
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap5==24.2
django-crispy-forms==2.2
django-import-export==4.1.1
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
msvc_runtime==14.40.33807
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prettier==0.0.7
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
